Divisão alterada para double

+++ Java2_poo_atp30/Calculadora.java
@@ -12,7 +12,7 @@ public class Calculadora {
         return n1 - n2;
     }
 
-    public int divisao(int n1, int n2) {
+    public double divisao(Double n1, Double n2) {
         return n1 / n2;
     }
